feat(platform): ãResponse ì 금액 자료형 수정 및 SecurityConfig.java 주석 ì추가

+++ platform/backend/src/main/java/com/woorifisan/platform/bank/dto/BalanceInquiryResponse.java
@@ -6,13 +6,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class BalanceInquiryResponse {
 
-    private Integer balance;
+    private BigDecimal balance;
     private String status;
 
 }
