+++ platform/backend/src/main/java/com/woorifisan/platform/bank/dto/TransactionHistoryDto.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -15,8 +17,8 @@ public class TransactionHistoryDto {
     private String txId;
     private String txDate;
     private String txType;
-    private Integer amount;
-    private Integer balance;
+    private BigDecimal amount;
+    private BigDecimal balance;
     private String counterpartName;
     private String description;
